refactor(datasource/docker): Consolidate files into one
